Ejercicio 04 - Kanban/VPS/AUTENTICACION-USUARIOS/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
cryptography